Implement new signature encoding and get rid of abi encoding in verifier/aggregator [CCIP-7253] (#115)
While working on this we discovered a bug with on-chain. To get around it we're not using mock reciever and using an EOA address instead so it doesn't go through router -> aggregator. Doing so we also changed executor to add the commitOfframp address manually instead of getting it from the receiver.