Add HeadTracker interface (#233)
This PR adds HeadTracker interface to get ready for implementing the reorg algorithm. Based on logpoller and headtracker team it's essential to use headtracker rather than the client directly. Using HeadTracker is more robust and resilient and that's how LogPoller detects reorgs and finality violations as well.

This also simplified the SourceReader interface and is making 1 call to get both latest and finalized blocks which reduced overall latency.

EvmSourceReader Implements both the SourceReader and HeadTracker interfaces