executor: fix protocol.UnknownAddress usage
The usage of the protocol.UnknownAddress was quite buggy. We used
.String() in some places and string() in others. It was really hard to
follow what's going on in the logs since the printed addresses were 2-3x
longer than they should be due to incorrect UnknownAddress
initialization.

Also cleaned up orderCCVData and added a lot of unit test cases.

This was discovered when adding test cases for multiple verifiers in
a message, which have been added to the test suite.