Perform aggregation of signature hash
Prior to this commit the aggregation was done over the messageID only. Up until recently this was fine because the messageID was the only piece of data that verifier attested to. However we recently add version to the blob data, which is not captured by the messageID hash. To enforce quorum on the actual data that verifier attest we changed the aggregation key to be the signature hash instead of the messageID