build/devenv/tests/e2e: add indexer down chaos test
Add a chaos test that brings down the indexer for a period of time where
it is needed to index verifier results, and bring it back up. The
indexer is expected to recover these verifier results and the executor
is expected to execute the message.

While writing the test a null pointer exception was found, so fixed that
by checking whether the response is null or not.