build/devenv/tests/e2e: add indexer down chaos test (#503)
* build/devenv/tests/e2e: add indexer down chaos test

Add a chaos test that brings down the indexer for a period of time where
it is needed to index verifier results, and bring it back up. The
indexer is expected to recover these verifier results and the executor
is expected to execute the message.

While writing the test a null pointer exception was found, so fixed that
by checking whether the response is null or not.

* run indexer chaos in a separate job

* timeouts per test in matrix

* add indexer restart policy

* try the always restart policy

* ping indexer db on startup

* try on-failure

* retry with sleeps in main.go

* combine indexer chaos with other tests to reduce ci time

* dump logs for timed out cl test